feat(application): extract staticAdapter and routedAdapter base helpers
Move all shared adapter logic (index.html reading, SEO injection, static
route file writing, SSG pre-render) from githubPagesAdapter into two
exported factory functions in adapters/adapter.mts:

- staticAdapter(definition) -- for file-based hosts (S3, GitHub Pages, ...);
  writes a catch-all fallback file (default 404.html) + all shared logic.
- routedAdapter(definition) -- for server-based hosts (Fastify, Express, ...);
  writes routes.json (base + staticRoutes + fallback) instead of a fallback
  file so the server can drive routing from the manifest.

Both expose AdapterContext (outputDirectory, indexHtml, config, manifestApi,
seoApi) to their setup() hook, allowing bespoke adapters to generate
host-specific artifacts from the same data the base adapter uses.

githubPagesAdapter() shrinks from 135 lines to ~20 by delegating to
staticAdapter and only handling the GitHub-specific .nojekyll write.
